LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/claim-shared-worker-fetch.https-expected.txt


PASS fetch() in SharedWorker should be intercepted after the client is claimed.
